Create 3. Voter checking
Write a program that has a class person storing name and date of birth (DOB) of a person. The program should subtract the DOB from today’s date to find out whether a person is eligible to vote or not using the constructor (A person is eligible to vote if age is
greater than or equal to 18).